requirements.txt
fastapi
uvicorn
python-multipart
numpy
opencv-python-headless
scapy
cryptography
pydantic
pycryptodome